backend/requirements.existing.lock.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==2.0.3
openai>=1.40.0,<2.0.0
google-generativeai==0.3.1
langgraph>=0.1.8,<0.2.0
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
anthropic==0.7.7
python-dotenv==1.0.0
httpx==0.24.1
numpy==1.26.2
scikit-learn==1.3.2
razorpay==1.3.0
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
PyPDF2==3.0.1
